Fixed several issues with the async query provider and missing results due to the stored queries not being arranged properly

+++ SanteDB.Core.Api/Data/RepositoryEntitySource.cs
@@ -64,7 +64,7 @@ namespace SanteDB.Core.Data
         /// <summary>
         /// Get versioned relationships for the object
         /// </summary>
-        public IEnumerable<TObject> GetRelations<TObject>(Guid? sourceKey, decimal? sourceVersionSequence) where TObject : IdentifiedData, IVersionedAssociation, new()
+        public IEnumerable<TObject> GetRelations<TObject>(Guid? sourceKey, int? sourceVersionSequence) where TObject : IdentifiedData, IVersionedAssociation, new()
         {
             return this.Query<TObject>(o => o.SourceEntityKey == sourceKey).ToList();
         }
